NIRCam coron off-axis PSF sums
Make sure the NIRCam coronagraphic off-axis PSFs are correctly attenuated by the coronagraphic mask. The total sum of an arbitrarily-positioned PSF  follows a simple scaling function based on the mask transmission at that point. Calculate on-axis and way off-axis PSFs and store sums/max values for later usage and scaling.